chore: remove old keyboards module
