+++ bot/handlers/start.py
@@ -1,7 +1,7 @@
 from telegram import Update
 from telegram.ext import ContextTypes
 
-from ..keyboards import main_menu
+from ..keyboards.constants import main_menu
 
 async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
     await update.message.reply_text(
