+++ bot/keyboards/__init__.py
@@ -1,0 +1,7 @@
+from .constants import *  # noqa: F401,F403
+from .builders import *  # noqa: F401,F403
+
+from .constants import __all__ as _constants_all
+from .builders import __all__ as _builders_all
+
+__all__ = [*_constants_all, *_builders_all]
